src/MCP/operations/requirements.txt
httpx
uvicorn
mcp==1.13.0
python-dotenv==1.0.0
azure-core==1.30.2
azure-identity==1.17.1
azure-cosmos
aiohttp
numpy
pandas
pyarrow


